1. Created the toolbar and added the test case 2. Moved the file to the correct directory 3. query param made rememberSavaeable
