+++ app/src/androidTest/java/com/ksale/pokemonsearch/SearchBarTest.kt
@@ -2,11 +2,12 @@ package com.ksale.pokemonsearch
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.ui.test.*
+import androidx.compose.ui.test.assertIsDisplayed
+import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.createComposeRule
-import com.ksale.pokemonsearch.ui.PokemonSearchScreen
+import androidx.compose.ui.test.onNodeWithContentDescription
+import androidx.compose.ui.test.performClick
 import com.ksale.pokemonsearch.ui.SearchBar
-import com.ksale.pokemonsearch.viewmodels.UIState
 import org.junit.Rule
 import org.junit.Test
