+++ app/src/main/java/com/ksale/pokemonsearch/MainActivity.kt
@@ -3,7 +3,7 @@ package com.ksale.pokemonsearch
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.ksale.pokemonsearch.ui.theme.PokemonSearchApp
+import com.ksale.pokemonsearch.ui.PokemonSearchApp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
