+++ app/src/main/java/com/ksale/pokemonsearch/ui/PokemonSearchApp.kt
@@ -1,11 +1,36 @@
-package com.ksale.pokemonsearch.ui.theme
+package com.ksale.pokemonsearch.ui
 
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
-import com.ksale.pokemonsearch.ui.PokemonSearchScreen
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.ksale.pokemonsearch.ui.theme.PokemonSearchTheme
 
 @Composable
 fun PokemonSearchApp() {
     PokemonSearchTheme {
-        PokemonSearchScreen()
+        Scaffold(
+            topBar = {
+                TopAppBar(
+                    elevation = 8.dp
+                ) {
+                    Text(
+                        text = "Pokemon Search",
+                        modifier = Modifier.padding(start = 16.dp),
+                        fontWeight = FontWeight.Bold,
+                        color = Color.White,
+                        fontSize = 22.sp
+                    )
+                }
+            }
+        ) {
+            PokemonSearchScreen()
+        }
     }
 }
